requirements.txt
accelerate==1.11.0
aiofiles==25.1.0
annotated-doc==0.0.3
annotated-types==0.7.0
anyio==4.11.0
backoff==2.2.1
beautifulsoup4==4.14.2
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
emoji==2.15.0
faiss-cpu==1.12.0
fastapi==0.121.0
filelock==3.20.0
filetype==1.2.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.10.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jiter==0.11.1
joblib==1.5.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==1.0.3
langchain-core==1.0.3
langdetect==1.0.9
langgraph==1.0.2
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.41
lxml==6.0.2
MarkupSafe==3.0.3
marshmallow==3.26.1
matplotlib==3.10.7
ml_dtypes==0.5.3
mpmath==1.3.0
mypy_extensions==1.1.0
networkx==3.5
nltk==3.9.2
numpy==2.2.6
olefile==0.47
onnx==1.19.1
onnxruntime==1.23.2
openai==2.7.1
opencv-python==4.12.0.88
orjson==3.11.4
ormsgpack==1.12.0
packaging==25.0
pandas==2.3.3
pdfminer.six==20250506
pi_heif==1.1.1
pillow==12.0.0
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
protobuf==6.33.0
psutil==7.1.3
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyMuPDF==1.26.5
pyparsing==3.2.5
pypdf==6.1.3
pypdfium2==5.0.0
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-iso639==2025.2.18
python-magic==0.4.27
python-multipart==0.0.20
python-oxmsg==0.0.2
pytz==2025.2
PyYAML==6.0.3
RapidFuzz==3.14.3
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
safetensors==0.6.2
scipy==1.16.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.8
starlette==0.49.3
sympy==1.14.0
tenacity==9.1.2
timm==1.0.21
tokenizers==0.22.1
torch==2.9.0
torchvision==0.24.0
tqdm==4.67.1
transformers==4.57.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
unstructured==0.18.15
unstructured-client==0.42.3
unstructured-inference==1.0.5
urllib3==2.5.0
uvicorn==0.38.0
webencodings==0.5.1
wrapt==2.0.0
xxhash==3.6.0
zstandard==0.25.0
